Move update plan tests to their own file
